backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
google-cloud-aiplatform
tqdm
scikit-learn
transformers
beautifulsoup4
python-multipart
aiofiles
python-dotenv
google-generativeai